modularize prompt components
