+++ packages/datasets/src/treeGeneration/databaseNlQueries/generateUseCases.ts
@@ -7,6 +7,7 @@ import {
   DatabaseUseCaseSchema,
   DatabaseUserNode,
 } from "./nodeTypes";
+import { makePromptDbInfo, makePromptDbUserInfo } from "./makePromptComponents";
 
 const useCaseSystemPrompt = `You are an expert user researcher who understands how different professionals use database information in their roles. Given a database user profile, generate realistic use cases that describe what information they need to retrieve from a database and why.
 
@@ -17,6 +18,7 @@ For each use case:
 - Emphasize the business context and purpose behind each information need
 - Consider both routine information needs and occasional specialized requirements
 - Include a balance of "simple", "moderate", and "complex" queries, as appropriate for a given user. Some users may have different balances, depending on their role.
+- ONLY make use cases that are supported by information in the given database.
 
 Generate no more than 10 distinct use cases that reflect different information needs relevant to their responsibilities.`;
 
@@ -40,22 +42,13 @@ export const generateDatabaseUseCases = makeGenerateChildrenWithOpenAi<
   }) => {
     const message = `Generate information retrieval use cases for the following user:
     
-Name: ${user.name}
-Job Title: ${user.jobTitle}
-Department: ${user.department}
-Description: ${user.description}
-Expertise: ${user.expertise.join(", ")}
-Years of Experience: ${user.yearsOfExperience}
+${makePromptDbUserInfo(user)}
 
 For a database with the following information:
-Name: ${databaseInfo.name}
-Description: ${databaseInfo.description}
-Collections:
-${databaseInfo.schema
-  .map((c) => ` - Name: ${c.name}, Description: ${c.description}`)
-  .join("\n")}
 
-Based on this profile, what are the most realistic and specific information needs this person would have when working with a movie database?`;
+${makePromptDbInfo(databaseInfo)}
+
+Based on this profile, what are the most realistic and specific information needs this person would have when working with the database?`;
 
     return [
       { role: "system", content: useCaseSystemPrompt },
